Split out search services
