+++ src/apps/search/services/index.js
@@ -1,0 +1,13 @@
+const search = require('./search')
+const searchCompanies = require('./search-companies')
+const searchForeignCompanies = require('./search-foreign-companies')
+const searchInvestments = require('./search-investments')
+const searchLimitedCompanies = require('./search-limited-companies')
+
+module.exports = {
+  search,
+  searchCompanies,
+  searchForeignCompanies,
+  searchInvestments,
+  searchLimitedCompanies,
+}
