+++ src/apps/search/services/search-foreign-companies.js
@@ -1,0 +1,11 @@
+const { assign } = require('lodash')
+
+const searchCompanies = require('./search-companies')
+
+function searchForeignCompanies (options) {
+  const optionsUkBasedFalse = assign({}, options, { isUkBased: false })
+
+  return searchCompanies(optionsUkBasedFalse)
+}
+
+module.exports = searchForeignCompanies
